+++ shopify/packages/ajax-cart-api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @brytdesigns/shopify-ajax-cart-api
 
+## 1.0.6
+
+### Patch Changes
+
+- Api now separates public and private attributes
+
 ## 1.0.5
 
 ### Patch Changes